+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 2.3.3
 - drop support of old android versions
+- use Square.OkHttp3.Android instead of Square.OkHttp3.JVM
+- remove Square.OkHtt3.JavaNetCookieJar (incompatible with Square.OkHttp3.Android)
 
 ## 2.3.2
 - vs : 17.14.12